refactoring bug fixes almost all rules work now
