+++ Lab1/CSharpFormatter.py
@@ -29,59 +29,23 @@ class CSharpFormatter:
         AnonymousTypeRule(),
     ]
 
-    def remove_double_whitespaces(self, tokens):
-        pos = 0
-        while pos < len(tokens)-1:
-            if tokens[pos][0] in whitespace_tokens+[Token.Newline] and tokens[pos+1][0] in whitespace_tokens:
-                del tokens[pos+1]
-            elif tokens[pos][0] in whitespace_tokens and tokens[pos+1][0] == Token.Newline:
-                del tokens[pos]
-            else:
-                pos+=1
-        return tokens
-
-    def split_whitespaces(self, tokens):
-        pos=0
-        new_tokens = []
-        whitespaces = [' ']
-        while pos < len(tokens):
-            if tokens[pos][0] == Token.Newline:
-                if pos+1 < len(tokens) and tokens[pos+1][0] == Token.Newline:
-                    whitespaces[-1]='\n'
-                    new_tokens.append((Token.EmptyLine, ""))
-                    whitespaces.append('\n')
-                    pos+=2
-                else:
-                    pos+=1
-            elif tokens[pos][0] not in whitespace_tokens:
-                new_tokens.append(tokens[pos])
-                whitespaces.append(' ')
-                pos+=1
-            else:
-                whitespaces[-1]=' '
-                pos+=1
-        return new_tokens, whitespaces
-
     def beautify(self, lexer):
         all_tokens=[]
         token, str = lexer.nextToken()
-
         while token != Token.EndOfInput:
             all_tokens.append((token, str))
             token, str = lexer.nextToken()
 
-        tokens, whitespaces = self.split_whitespaces(self.remove_double_whitespaces(all_tokens))
-        code = Code(tokens, whitespaces)
-        for rule in self.rules:
-            rule.code = code
+        token_processor = CSharpTokenProcessor(all_tokens)
+        token_processor.check_single_tokens()
 
-        self.check_single_tokens(code)
+        for rule in self.rules:
+            rule.set_token_processor(token_processor)
 
         pos = 0
         stack = []
-        while pos < len(tokens):
-            print(tokens[pos][0])
-            stack.insert(0, ASTNode(tokens[pos][0], pos, pos+1))
+        while pos < len(token_processor.tokens):
+            stack.insert(0, ASTNode(token_processor.get_token(pos), pos, pos+1))
             reduced = True
             while reduced:
                 reduced = False
@@ -93,110 +57,54 @@ class CSharpFormatter:
                         break
             pos+=1
 
-        return code.get_str()
-
-    def check_single_tokens(self, code):
-        for pos in range(len(code.tokens)):
-            token = code.get_token(pos)
-            if token == Tokens[',']:
-                code.set_space_before(pos, ' ' if csharp_space_before_comma else '')
-                code.set_space_after(pos, ' ' if csharp_space_after_comma else '')
-            elif token == Tokens['.']:
-                code.set_space_before(pos, ' ' if csharp_space_before_dot else '')
-                code.set_space_after(pos, ' ' if csharp_space_after_dot else '')
-            elif token in [Tokens[i] for i in BinaryOperators]:
-                code.set_space_before(pos, ' ' if csharp_space_around_binary_operators else '')
-                code.set_space_after(pos, ' ' if csharp_space_around_binary_operators else '')
-            elif token == Tokens['catch']:
-                code.set_space_before(pos, '\n' if csharp_new_line_before_catch else ' ')
-            elif token == Tokens['finally']:
-                code.set_space_before(pos, '\n' if csharp_new_line_before_finally else ' ')
-            elif token == Tokens['else']:
-                code.set_space_before(pos, '\n' if csharp_new_line_before_else else ' ')
-            elif token == Tokens['{']:
-                code.set_space_before(pos, '\n' if csharp_new_line_before_open_brace else ' ')
-            elif token == Tokens[';']:
-                code.set_space_before(pos, '')
-                code.set_space_after(pos, '\n')
-
-
-
-def parse_template(path):
-    template = None
-    with open(path) as f:
-        f.read()
-    return template
-
-
-if __name__ == "__main__":
-    formatter = CSharpFormatter()
-
-    parser = argparse.ArgumentParser(description="Specify input and output!")
-    parser.add_argument("--input_path", help="path to format", required=True)
-    parser.add_argument("--output_path", help="path to save result", required=True)
-
-    group = parser.add_mutually_exclusive_group(required=True)
-    group.add_argument("-f", "--file", help="file", action='store_true')
-    group.add_argument("-d", "--directory", help="directory", action='store_true')
-    group.add_argument("-p", "--project", help="project", action='store_true')
-
-    parser.add_argument("--template", help="path to template file")
-    args = parser.parse_args()
-
-    template_path = None if not args.template else args.template
-    template = parse_template(template_path)
-
-    if args.file:
-        with open(args.input_path) as fi:
-            str = fi.read()
-            beautified_str = formatter.beautify(str, template)
-            out_file = args.output_path
-            if not os.path.exists(os.path.dirname(out_file)):
-                try:
-                    os.makedirs(os.path.dirname(out_file))
-                except OSError as exc: # Guard against race condition
-                    if exc.errno != errno.EEXIST:
-                        raise
-            with open(out_file, 'w') as fo:
-                fo.write(beautified_str)
-    elif args.directory:
-        for dir, _, files in os.walk(args.input_path):
-            rel_dir = os.path.relpath(dir, args.input_path)
-            for file in fnmatch.filter(files, "*.cs"):
-                rel_file = os.path.join(rel_dir, file)
-                with open(os.path.join(dir, file)) as f:
-                    str = f.read()
-                    beautified_str = formatter.beautify(str, template)
-                    out_file = os.path.join(args.output_path, rel_file)
-                    if not os.path.exists(os.path.dirname(out_file)):
-                        try:
-                            os.makedirs(os.path.dirname(out_file))
-                        except OSError as exc: # Guard against race condition
-                            if exc.errno != errno.EEXIST:
-                                raise
-                    with open(out_file, 'w') as fo:
-                        fo.write(beautified_str)
-
-lexer = CSharpLexer("""namespace System.Main.Complex {
-public class IAbstractFactory<T>
-{
-    void main<T> ( int a, S b ) {
-        main();
-    }
-}
-}""")
-
+        return token_processor.get_str()
 
 lexer = CSharpLexer("""/*
+ /*
  * C# Program to Perform Bubble Sort
  */
 
-if (a[i] > a[i + 1])
+using System;
+class bubblesort
 {
-    t = a[i + 1];
-    a[i + 1] = a[i];
-    a[i] = t;
-}
+        static void Main(string[] args)
+        {
+            int[] a = { 3, 2, 5, 4, 1 };
+            int t;
+            Console.WriteLine("The Array is : ");
+            for (int i = 0; i < a.Length; i++)
+            {
+                Console.WriteLine(a[i]);
+            }
+            for (int j = 0; j <= a.Length - 2; j++)
+            {
+                for (int i = 0; i <= a.Length - 2; i++)
+                {
+                    if (a[i] > a[i + 1])
+                    {
+                        t = a[i + 1];
+                        a[i + 1] = a[i];
+                        a[i] = t;
+
+                        switch () {
+                        case 1:
+                        a = 1;
+                        break;
+                        case 2: {a = 1;}break;
+                        default: { a = 0;  }break;
+                        }
+                    }
+                }
+            }
+            Console.WriteLine("The Sorted Array :");
+
+            foreach (int aray in a)
+                Console.Write(aray + " ");
+
+            Console.ReadLine();
+        }
+    }
 """)
-s = CSharpFormatter().beautify(lexer)
+formatter = CSharpFormatter()
+s = formatter.beautify(lexer)
 print(s)
